Update sys_selection upon disconnect
sys_selection was not reset after a failed connection attempt. This caused
the device to search for LTE, which did of course not work (I have to be on UMTS
to be handed over to LTE). Also, only update NAS state when previous state is
correct (RESET in send, SENT_SYS on handle).

This commit also adds reading the current band that the modem uses upon timeout.
The values are defined in the QMI NAS spec.