Use of sys_selection is more robust
Updating system selection after a connection was established caused different
types of problems. For example, the MF821D got stuck thinking it was on LTE and
did not bother searching for networks, even though it was actually out of
service. It turns out that it is sufficient to only set the system_selection