Support for connecting to UMTS/GSM first
Some operators do not allow UE to connect directly to LTE, a connection to UMTS
or GSM has to be established first. This commit enables support for temporarily
locking the modem to UMTS/GSM until a successful connection is made.

The mode is named after the first operator where this was seen. Be aware that
updating the system selection mode after connect can make the modem behave
strange if it is not power-cycled between connects.