[FIX] Remove Netcom hack
I initially thought that on some ISPs, Norwegian Netcom included, that UE had to
attach and connect to UMTS before connection to LTE was possible. This is wrong,
it is only the APN that has to be correct (and on some modems it HAS to be set
using AT+ZSNT).